requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
SQLAlchemy==2.0.36